Rename UnifiedBody to Body, make motion readonly
Fold UnifiedBody into the canonical Body class and delete the transitional
re-export shim. Make `motion` readonly since the engine doesn't support
flipping motion role at runtime. Fix angularVelocity init to use
`!== undefined` so an explicit 0 is respected. Update physics/boat docs
to match the unified body + systems architecture.
