+++ src/game/constraints/TerrainFloorConstraint.ts
@@ -10,7 +10,7 @@
 
 import { BaseEntity } from "../../core/entity/BaseEntity";
 import { on } from "../../core/entity/handler";
-import type { UnifiedBody } from "../../core/physics/body/UnifiedBody";
+import type { Body } from "../../core/physics/body/Body";
 import { V, V2d } from "../../core/Vector";
 import { TerrainQuery } from "../world/terrain/TerrainQuery";
 import { WaterQuery } from "../world/water/WaterQuery";
@@ -21,14 +21,14 @@ export interface TerrainFloorOptions {
 }
 
 export class TerrainFloorConstraint extends BaseEntity {
-  private constrainedBodies: UnifiedBody[];
+  private constrainedBodies: Body[];
   private queryPoints: V2d[];
   private terrainQuery: TerrainQuery;
   private waterQuery: WaterQuery;
   private floorFriction: number;
 
   constructor(
-    bodies: UnifiedBody[] | readonly UnifiedBody[],
+    bodies: Body[] | readonly Body[],
     options?: TerrainFloorOptions,
   ) {
     super();
